refactor(bookmarks): fix filenames and imports
- renames for consistency
- proper exports/imports through index files

SVA-62
