+++ src/hooks/Bookmarks.ts
@@ -1,6 +1,6 @@
 import { useCallback, useEffect, useState } from 'react';
+import { BookmarkList, getBookmarkedItems } from '../helpers';
 
-import { BookmarkList, getBookmarkedItems } from '../helpers/bookmarkHelpers';
 
 export const useBookmarks = (itemType?: string, category?: number) => {
   const [bookmarks, setBookmarks] = useState<BookmarkList | string[] | undefined>();
